Add Basic CI/CD Workflow (#37)
* Add basic CI/CD workflow

* Update CI workflow to target dev branch

* update workflow

* limit matrix (since we use docker)

* fix ci/cd issues

* update ci to load HF_TOKEN

* load environment variable for testing

---------
