+++ .github/workflows/ci.yml
@@ -1,0 +1,34 @@
+name: CI
+
+on:
+  push:
+    branches: [ dev, main ]
+  pull_request:
+    branches: [ dev, main ]
+
+jobs:
+  build:
+    runs-on: rtx3090x1
+
+    steps:
+    - uses: actions/checkout@v2
+
+    - name: Set up Docker Buildx
+      uses: docker/setup-buildx-action@v3
+
+    - name: Build Docker image
+      run: |
+        bash docker/build_image.sh dev docker no-upload
+
+    - name: Run tests in Docker container
+      env:
+        HF_TOKEN: ${{ secrets.HF_TOKEN }}
+      run: |
+        # Run your test suite inside the Docker container
+        # Replace this with your actual test command
+        docker run --gpus all --ipc=host --ulimit memlock=-1 --ulimit stack=67108864 -e HF_TOKEN=$HF_TOKEN ghcr.io/xiaozheyao/scratchpad:devdev-x86_64 python -m pytest -s tests/
+
+    - name: Docker image cleanup
+      if: always()
+      run: |
+        docker image prune -f
